Restructure frontend directory, including lifting out dialects
