+++ xdsl/frontend/pyast/context.py
@@ -6,7 +6,7 @@ from sys import _getframe  # pyright: ignore[reportPrivateUsage]
 from typing import Any
 
 from xdsl.frontend.pyast.program import FrontendProgram
-from xdsl.frontend.pyast.python_code_check import PythonCodeCheck
+from xdsl.frontend.pyast.utils.python_code_check import PythonCodeCheck
 
 
 @dataclass
