+++ xdsl/transforms/__init__.py
@@ -289,7 +289,7 @@ def get_all_passes() -> dict[str, Callable[[], type[ModulePass]]]:
         return eqsat_extract.EqsatExtractPass
 
     def get_frontend_desymrefy():
-        from xdsl.frontend.pyast.passes.desymref import FrontendDesymrefyPass
+        from xdsl.transforms.desymref import FrontendDesymrefyPass
 
         return FrontendDesymrefyPass
